# docs/openapi.yaml
openapi: 3.0.3
info:
  title: SafeShare API
  description: |
    SafeShare is a secure file sharing service with support for large file uploads, 
    password protection, expiration controls, and enterprise features.
    
    ## Authentication
    
    SafeShare supports two authentication methods:
    
    ### Session Cookie Authentication
    Traditional web session using `user_session` cookie. Obtained via `/api/auth/login`.
    
    ### API Token Authentication (Bearer)
    For SDKs, CLIs, and automation. Include token in Authorization header:
    ```
    Authorization: Bearer safeshare_<token>
    ```
    
    API tokens support granular scopes: `upload`, `download`, `manage`, `admin`.
    
    ## Rate Limiting
    
    Rate limits are applied per IP address:
    - Uploads: Configurable (default: 10/hour)
    - Downloads: Configurable (default: 50/hour)
    - Login attempts: 5 per 15 minutes
    - Chunked upload: `/init` at the upload limit, `/chunk` and `/complete` at
      10x it, `/status` at 600x it (never under 6000/hour)

    When rate limited, endpoints return `429 Too Many Requests`.

  version: 1.12.0
  contact:
    name: SafeShare Support
    url: https://github.com/fjmerc/safeshare
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080
    description: Development server
  - url: https://yourdomain.com
    description: Production server
...
paths:
...
  /api/auth/user:
    get:
      tags:
        - Authentication
      summary: Get current user
      description: Retrieve information about the currently authenticated user.
      operationId: getCurrentUser
      security:
        - sessionAuth: []
        - bearerAuth: []
      responses:
        '200':
          description: Current user info
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /admin/api/users:
    get:
      tags:
        - Admin
      summary: List users
      description: Retrieve all users in the system.
      operationId: listUsers
      security:
        - adminAuth: []
      responses:
        '200':
          description: List of users
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserListResponse'
...
  /admin/api/users/{id}:
    put:
      tags:
        - Admin
      summary: Update user
      description: Update an existing user account.
      operationId: updateUser
      security:
        - adminAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: User ID
        - $ref: '#/components/parameters/CSRFToken'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
      responses:
        '200':
          description: User updated successfully
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      tags:
        - Admin
      summary: Delete user
      description: Delete a user account.
      operationId: deleteUser
      security:
        - adminAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: User ID
        - $ref: '#/components/parameters/CSRFToken'
      responses:
        '200':
          description: User deleted successfully
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: user_session
      description: User session cookie obtained via /api/auth/login

    bearerAuth:
      type: http
      scheme: bearer
      description: |
        API token authentication. Token format: `safeshare_<64 hex chars>`
        
        Example: `Authorization: Bearer safeshare_a1b2c3d4...`

    adminAuth:
      type: apiKey
      in: cookie
      name: admin_session
      description: Admin session cookie obtained via /admin/api/login
  parameters:
    CSRFToken:
      name: X-CSRF-Token
      in: header
      required: true
      schema:
        type: string
      description: CSRF token from csrf_token cookie
...
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Human-readable error message
        code:
          type: string
          description: Machine-readable error code
      required:
        - error

    # ==================== Health Schemas ====================
...
    UserResponse:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        email:
          type: string
          format: email
        role:
          type: string
          enum: [user, admin]
        require_password_change:
          type: boolean
...
    UserListResponse:
      type: object
      properties:
        users:
          type: array
          items:
            $ref: '#/components/schemas/AdminUserInfo'
    AdminUserInfo:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        email:
          type: string
          format: email
        role:
          type: string
          enum: [user, admin]
        is_active:
          type: boolean
        created_at:
          type: string
          format: date-time
        last_login_at:
          type: string
          format: date-time
          nullable: true
...
    UpdateUserRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        role:
          type: string
          enum: [user, admin]

    # ==================== Webhook Schemas ====================
...